requirements.txt
openai
faster-whisper
sounddevice
scipy
numpy
gtts
python-dotenv
